+++ examples/standard-use/open-ai-ask.ts
@@ -6,15 +6,15 @@ import { JorEl } from "../../src";
 config({ path: "../../.env" });
 
 const main = async () => {
-  // Create instance
+  // Create instance with pre-initialized OpenAI provider (and default models)
   const jorEl = new JorEl({
     openAI: { apiKey: process.env.OPENAI_API_KEY },
   });
 
-  // Will return a string
-  const response = await jorEl.ask("What is the capital of France?");
+  const response = await jorEl.ask("What is the capital of France, in one word?");
 
   console.log(response);
+  // Paris
 };
 
 void main();